refactor(BA-5978): move BackendAIModel from exception to types
Belongs alongside other shared base types in ``common.types`` rather
than mixed with the error hierarchy. ``ModelValidationFailed`` stays
in ``exception.py`` and is imported by ``types.py`` for the
auto-conversion path.
