+++ src/ai/backend/common/types.py
@@ -5,6 +5,7 @@ import dataclasses
 import enum
 import ipaddress
 import itertools
+import logging
 import math
 import numbers
 import textwrap
@@ -49,14 +50,18 @@ from pydantic import (
     Field,
     PlainValidator,
     TypeAdapter,
+    ValidationError,
 )
 from redis.asyncio import Redis
+
+from ai.backend.logging import BraceStyleAdapter
 
 from .defs import UNKNOWN_CONTAINER_ID, RedisRole
 from .exception import (
     GenericNotImplementedError,
     InvalidIpAddressValue,
     InvalidResourceSlotQuantity,
+    ModelValidationFailed,
 )
 
 # Deprecated re-export: new code should import ``ImageID`` from
@@ -81,6 +86,7 @@ __all__ = (
     "AutoPullBehavior",
     "AutoScalingMetricComparator",
     "AutoScalingMetricSource",
+    "BackendAIModel",
     "BinarySize",
     "CIStrEnum",
     "CIStrEnumTrafaret",
@@ -173,9 +179,62 @@ if TYPE_CHECKING:
     from .docker import ImageRef
 
 
+log = BraceStyleAdapter(logging.getLogger(__spec__.name))
+
+
 current_resource_slots: ContextVar[Mapping[SlotName, SlotTypes]] = ContextVar(
     "current_resource_slots"
 )
+
+
+class BackendAIModel(BaseModel):
+    """Project-wide Pydantic base for Backend.AI models.
+
+    Overrides ``model_validate`` / ``model_validate_json`` /
+    ``model_validate_strings`` so a ``ValidationError`` is auto-mapped
+    to :class:`ModelValidationFailed` (HTTP 400) carrying the structured
+    per-field error list. Call sites get a clean 4xx without repeating
+    ``try / except ValidationError`` at every site.
+
+    The raw Pydantic ``ValidationError`` is also logged at ``error``
+    level so operators can see the original message even when the
+    converted ``BackendAIError`` truncates or rephrases it.
+
+    Notes:
+
+    * Pydantic v2 routes nested validation through
+      ``__pydantic_validator__`` directly, not the classmethod, so this
+      override only affects explicit ``Model.model_validate(...)``
+      calls — nested models are unaffected.
+    * The ``__init__`` constructor and the compiled validator stay
+      untouched, so internal default-value construction
+      (``Model()`` / ``Model(field=...)``) still works exactly like
+      stock Pydantic.
+    """
+
+    @classmethod
+    def model_validate(cls, *args: Any, **kwargs: Any) -> Self:
+        try:
+            return super().model_validate(*args, **kwargs)
+        except ValidationError as e:
+            log.error("Pydantic validation failed for {}: {}", cls.__name__, e)
+            raise ModelValidationFailed.from_pydantic(e) from e
+
+    @classmethod
+    def model_validate_json(cls, *args: Any, **kwargs: Any) -> Self:
+        try:
+            return super().model_validate_json(*args, **kwargs)
+        except ValidationError as e:
+            log.error("Pydantic validation failed for {}: {}", cls.__name__, e)
+            raise ModelValidationFailed.from_pydantic(e) from e
+
+    @classmethod
+    def model_validate_strings(cls, *args: Any, **kwargs: Any) -> Self:
+        try:
+            return super().model_validate_strings(*args, **kwargs)
+        except ValidationError as e:
+            log.error("Pydantic validation failed for {}: {}", cls.__name__, e)
+            raise ModelValidationFailed.from_pydantic(e) from e
 
 
 class aobject:
