+++ src/ai/backend/manager/data/session/spec.py
@@ -23,12 +23,11 @@ from uuid import UUID
 import yarl
 from pydantic import ConfigDict, Field
 
-from ai.backend.common.exception import BackendAIModel
 from ai.backend.common.identifier.domain import DomainName
 from ai.backend.common.identifier.project import ProjectID
 from ai.backend.common.identifier.resource_group import ResourceGroupName
 from ai.backend.common.identifier.session import SessionID
-from ai.backend.common.types import AccessKey, SessionTypes, VFolderMount
+from ai.backend.common.types import AccessKey, BackendAIModel, SessionTypes, VFolderMount
 from ai.backend.manager.data.session.options import (
     InternalDataExtras,
     KernelExecutionSpec,
