+++ src/ai/backend/manager/models/scaling_group/row.py
@@ -30,10 +30,10 @@ from sqlalchemy.orm import (
 )
 from sqlalchemy.sql.expression import true
 
-from ai.backend.common.exception import BackendAIModel
 from ai.backend.common.identifier.project import ProjectID
 from ai.backend.common.types import (
     AgentSelectionStrategy,
+    BackendAIModel,
     PreemptionMode,
     PreemptionOrder,
     SessionTypes,
